Use the new components for the Patients page
