+++ frontend/src/shared/components/Heading.tsx
@@ -3,11 +3,10 @@ import { buildSentenceCaseString } from "../../utils/stringBuilders";
 
 // TODO: change this back to Title.tsx
 export function Heading({ children }: { children: string }) {
-  const heading = buildSentenceCaseString(children);
   return (
     <Row className="pagetitle">
       <Col>
-        <h1>{heading}</h1>
+        <h1>{buildSentenceCaseString(children)}</h1>
       </Col>
     </Row>
   );
